형변환 경고들 명시적으로 표기

+++ winAPI_2Dcopy/CEnemy.cpp
@@ -40,7 +40,7 @@ void CEnemy::setRandPos(int hSize)
 		y = hSize;
 		break;
 	}
-	setPos(fPoint(x, y));
+	setPos(fPoint((float)x, (float)y));
 }
 
 void CEnemy::setRandDir()
@@ -49,18 +49,18 @@ void CEnemy::setRandDir()
 	fPoint size = getSize();
 
 	if ((int)pos.x <= (int)size.x / 2)							// 위치 : 좌 -> xDir : 우
-		fvDir.x = rand() % 5;
+		fvDir.x = (float)(rand() % 5);
 	else if ((int)pos.x >= (int)WINSIZEX - (int)size.x / 2)		// 위치 : 우 -> xDir : 좌
-		fvDir.x = rand() % 5 - 4;
+		fvDir.x = (float)(rand() % 5 - 4);
 	else
-		fvDir.x = rand() % 9 - 4;
+		fvDir.x = (float)(rand() % 9 - 4);
 
 	if ((int)pos.y <= (int)size.x / 2)							// 위치 : 상 -> yDir : 하
-		fvDir.y = rand() % 5;
+		fvDir.y = (float)(rand() % 5);
 	else if ((int)pos.y >= (int)WINSIZEY - (int)size.x / 2)		// 위치 : 하 -> yDir : 상
-		fvDir.y = rand() % 5 - 4;
+		fvDir.y = (float)(rand() % 5 - 4);
 	else
-		fvDir.y = rand() % 9 - 4;
+		fvDir.y = (float)(rand() % 9 - 4);
 }
 
 void CEnemy::setDir(fVec2 vec)
